# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Chatter messaging app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  inviteCode: string;
  createdAt: string;
}

export interface RegisterInput {
  email: string;
  username: string;
  password: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Conversation {
  id: number;
  otherUser: User;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  unreadCount: number;
}

export interface ConversationInput {
  otherUserId: number;
}

export interface Message {
  id: number;
  conversationId: number;
  senderId: number;
  content: string;
  messageType: string;
  createdAt: string;
  isOwn: boolean;
}

export interface MessageInput {
  content: string;
  messageType?: string;
}

export interface ProfileUpdate {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface InviteLink {
  link: string;
  inviteCode: string;
  qrData: string;
}

export type SearchUsersParams = {
q: string;
};