# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Chatter messaging app API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: users
    description: User management
  - name: conversations
    description: Conversations/chats
  - name: messages
    description: Messages
  - name: profile
    description: Profile management
  - name: invite
    description: Invite friends

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by username
      parameters:
        - in: query
          name: q
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Users found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - in: path
          name: userId
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

  /conversations:
    get:
      operationId: getConversations
      tags: [conversations]
      summary: Get all conversations for current user
      responses:
        "200":
          description: Conversations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [conversations]
      summary: Create or get a conversation with a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /conversations/{conversationId}:
    get:
      operationId: getConversation
      tags: [conversations]
      summary: Get a conversation by ID
      parameters:
        - in: path
          name: conversationId
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
          description: Not found

  /conversations/{conversationId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - in: path
          name: conversationId
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - in: path
          name: conversationId
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /messages/{messageId}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - in: path
          name: messageId
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  /profile:
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /invite/link:
    get:
      operationId: getInviteLink
      tags: [invite]
      summary: Get shareable invite link for current user
      responses:
        "200":
          description: Invite link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteLink"

  /invite/accept/{inviteCode}:
    get:
      operationId: acceptInvite
      tags: [invite]
      summary: Get user info from invite code
      parameters:
        - in: path
          name: inviteCode
          schema:
            type: string
          required: true
      responses:
        "200":
          description: User from invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Invite not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        inviteCode:
          type: string
        createdAt:
          type: string
      required:
        - id
        - username
        - displayName
        - inviteCode
        - createdAt

    RegisterInput:
      type: object
      properties:
        email:
          type: string
        username:
          type: string
        password:
          type: string
      required:
        - email
        - username
        - password

    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token

    Conversation:
      type: object
      properties:
        id:
          type: integer
        otherUser:
          $ref: "#/components/schemas/User"
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
        unreadCount:
          type: integer
      required:
        - id
        - otherUser
        - unreadCount

    ConversationInput:
      type: object
      properties:
        otherUserId:
          type: integer
      required:
        - otherUserId

    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        content:
          type: string
        messageType:
          type: string
        createdAt:
          type: string
        isOwn:
          type: boolean
      required:
        - id
        - conversationId
        - senderId
        - content
        - messageType
        - createdAt
        - isOwn

    MessageInput:
      type: object
      properties:
        content:
          type: string
        messageType:
          type: string
      required:
        - content

    ProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string

    InviteLink:
      type: object
      properties:
        link:
          type: string
        inviteCode:
          type: string
        qrData:
          type: string
      required:
        - link
        - inviteCode
        - qrData