# lib/api-zod/src/generated/types/inviteLink.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Chatter messaging app API
 * OpenAPI spec version: 0.1.0
 */

export interface InviteLink {
  link: string;
  inviteCode: string;
  qrData: string;
}